Changed the color list in GameBoard to ColorSet.
Added a caching bitset to MutableGame and SimplifiedGame to use instead of always creating a new bitset for the nodes to be added.

Depending on how many neighbors there are, it is quicker to just check whether a color happens to be in the neighbor bitset for every single color.