Removed parallel A* algorithm.
This implementation was a holdover from when leaf processing was so slow that it was advantageous to use multiple threads to do it. This is no longer the case and testing showed that the sequential algorithm performs better than its parallel sibling.

If I can find a performant parallel algorithm, I may reimplement this behavior, but until then it is better to just use the sequential algorithm.