Created a complete optimal solutionset for floodtest with terminal-flood.
This is a nice milestone, as performance optimizations and the addition of memory saving schemes finally allowed to solve the complete floodtest dataset in a reasonable amount of time and with a maximum heapsize of 12G.

solutions optimal floodtest.txt now contains the solutions I computed with terminal-flood instead of what smack42 had provided with ColorFill. The score is thankfully the same, which I take as confirmation that both programs' implementation of the admissible A* algorithm can solve game boards without error.

And now to figure out how to do it with the b24c6n1000 dataset!