Removed BoardNode.DUMMY_NODE. The GameBoard lookup table is initialized with a nullable array instead which gets casted to a non-nullable array (a save operation in that context).
Changed the default value of queueMaxSizeCutoff to Int.MAX_VALUE, meaning not doing a queue cutoff is the default. Currently the queue cutoff value is the same in Util and Datasets, but I'm debating whether it makes sense to use different values there (Util is only used single-threaded, while Datasets can be used in a multi-threaded context with the corresponding higher memory demands)

Ran the b24c6n1000 dataset with astat_a and a custom queue cutoff of 2000000, the current potentially optimal solutionset now has a score of 31807.