Instead of sorting the nodes based on their priority, run Greedy on all of them and use the thus predicted solution length to sort the nodes with priority merely being the tie breaker. We then discard half the nodes instead of just a fifth as before.
This resulted in about double the runtime length for the b24c6n1000 dataset with 4 threads, but the resulting solutionset is now strictly better than astar_ias and in fact even 3 score points better than the previous potentially optimal solutionset with a score of 31804. I'd say this is a good exchange.

Changed the queue cutoff in Util to 2000000.