Turned most strategies from objects into classes.
Originally, all strategies were objects with side-effect free heuristic() implementations to facilitate easy usage of them in a multi-threaded context. Since I've given up multi-threaded implementations of an A* algorithm for the time being, it made sense to turn them into classes that keep fields for their internal state around so as to save on the amount of object creations. The result is slightly better performance of all solver algorithms, especially in cases of heavier memory usage (probably due to less work needed to be done by the GC).