Implemented a custom hashtable for the board state lookups used by AStar.
This implementation is heavily inspired by a similar hashtable implementation done by smack42 for ColorFill. The main difference is probably that I (in part) use fibonacci hashing. It's absolutely possible that fibonacci hashing is not the best possible choice for hashing, but I really like the simplicity of the algorithm and the linked website shows that it certainly isn't a bad choice either for this particular use case.

Testing showed that performance increased slightly for problemsets that haven't the highest memory workloads, but for memory heavy problemsets (basically all the -lessMemory cases) performance increased quite a bit. This leads me to believe that the straight performance of this hashtable isn't that much better than Java's HashMap, but this implementation is much more efficient with regards to memory usage and trumps HashMap by making the GC do less work.

For the record, I also tested a single array version of this hashtable where keys and values are put into the same internal array. I had hoped that the additional data locality might improve performance further, but it didn't. It actually performed worse than the double array version, possibly due to the additional memory demands from having to save the values as Longs instead of Shorts.