Implemented a more memory efficient way to store move lists.
With problemsets of relatively low GC pressure this seems to perform from about the same to minimally worse, but with problemsets of high GC pressure this brings pretty decent performance improvements due to the lowered memory requirements.

I want to note that I'm not 100% happy with this change, because I think it made the design of the implementing GameState type worse. Before, everything necessary to represent a Flood-It game state was part of the game state (see Game), now GameExternalMoveList has to be used together with MoveCollection to be fully functioning. It was still the right call to do this due to the lower memory requirements, but it is a less straight forward and clear design. This is why I'm keeping Game and MoveList around, as whenever the additional performance benefits are of no concern, Game is still the better choice to use.