Added the ability to ask for a hint by the computer on which move to make while playing the game.
Currently only uses AStarStrategies.INADMISSIBLE_FAST (astar_iaf) to compute the solution the hint is based on. It should be more than fast enough for all game board sizes and amount of colors that can be reasonably expected to be played while still giving pretty decent solutions. May allow to customize that in the future.