Turn Color into a value class.
I can finally do away with color caches and just use Color like a mostly normal class. This was always how I imagined the Color class to be used and I'm happy that it can finally be implemented this way.

Value classes are essentially a zero-cost abstraction as long as you are mindful of a couple pitfalls (basically the same sort of stuff you have to pay attention to with Int), so this change shouldn't cost any performance.